now with permissions tweak that allow for readonly users

+++ Sources/wiremand/Databases/DaemonDB/DaemonDB.swift
@@ -3,16 +3,11 @@ import Foundation
 import CLMDB
 import NIO
 import SwiftSMTP
+import SystemPackage
+import Logging
 
 extension Task:MDB_convertible {
     public init?(_ value: MDB_val) {
-        guard MemoryLayout<Self>.stride == value.mv_size else {
-            return nil
-        }
-        self = value.mv_data.bindMemory(to:Self.self, capacity:1).pointee
-    }
-    
-    public init?(noCopy value: MDB_val) {
         guard MemoryLayout<Self>.stride == value.mv_size else {
             return nil
         }
@@ -27,9 +22,33 @@ extension Task:MDB_convertible {
     }
 }
 
-actor DaemonDB {
+class DaemonDB {
+	fileprivate static func makeLogger() -> Logger {
+		var newLogger = Logger(label:"daemon-db")
+		#if DEBUG
+			newLogger.logLevel = .trace
+		#else
+			newLogger.logLevel = .info
+		#endif
+		return newLogger
+	}
+	internal static let logger = makeLogger()
+	
 	static func create(directory:URL, publicHTTPPort:UInt16, notify:Email.Contact) throws -> Environment {
-		let makeEnv = try Environment(path:directory.appendingPathComponent("daemon-dbi").path, flags:[.noSubDir], mapSize:75000000000, maxDBs:128, mode: [.ownerReadWriteExecute])
+		let makeEnvPath = directory.appendingPathComponent("daemon-dbi")
+		let makeEnvLockPath = directory.appendingPathComponent("daemon-dbi-lock")
+		
+		//verify that the files do not already exist
+		if access(makeEnvPath.path, F_OK) == 0 {
+			try FileManager.default.removeItem(at:makeEnvPath)
+			try FileDescriptor.open(makeEnvPath.path, .writeOnly, options:[.create], permissions:[.ownerReadWriteExecute, .groupReadWriteExecute, .otherRead, .otherExecute]).close()
+		}
+		if access(makeEnvLockPath.path, F_OK) == 0 {
+			try FileManager.default.removeItem(at:makeEnvLockPath)
+			try FileDescriptor.open(makeEnvLockPath.path, .writeOnly, options:[.create], permissions:[.ownerReadWriteExecute, .groupReadWriteExecute, .otherReadWriteExecute]).close()
+		}
+		let makeEnv = try Environment(path:makeEnvPath.path, flags:[.noSubDir], mapSize:75000000000, maxDBs:128, mode: [.ownerReadWriteExecute, .groupReadWriteExecute])
+		logger.info("lmdb environment successfully created")
         try makeEnv.transact(readOnly:false) { someTrans in
             let metadata = try makeEnv.openDatabase(named:Databases.metadata.rawValue, flags:[.create], tx:someTrans)
             _ = try makeEnv.openDatabase(named:Databases.scheduleTasks.rawValue, flags:[.create], tx:someTrans)
@@ -52,6 +71,8 @@ actor DaemonDB {
     enum Error:Swift.Error {
         case daemonAlreadyRunning
         case pidExclusivityNotClaimed
+		case databaseNotCreated
+		case databaseNotReadable
     }
     enum Metadatas:String {
         case daemonRunningPID = "_daemonRunningPID" //pid_t
@@ -64,6 +85,9 @@ actor DaemonDB {
     
 	let loopGroup = MultiThreadedEventLoopGroup(numberOfThreads:System.coreCount)
 	
+	let readOnly:Bool
+	let running:Bool
+	
     let env:Environment
     let metadata:Database
     let scheduledTasks:Database
@@ -72,32 +96,64 @@ actor DaemonDB {
 	let notifyDB:Database
 	
     let wireguardDatabase:WireguardDatabase
-	let printerDatabase:PrintDB
+	let printerDatabase:PrintDB?
 	let ipdb:IPDatabase
 	
-    init(directory:URL, running:Bool = true) throws {
-		let makeEnv = try Environment(path:directory.appendingPathComponent("daemon-dbi").path, flags:[.noSubDir, .noSync, .noReadAhead], mapSize:75000000000, maxDBs:128)
-		try makeEnv.readerCheck()
-		let dbs = try makeEnv.transact(readOnly:false) { someTrans -> [Database] in
-            let metadataDB = try makeEnv.openDatabase(named:Databases.metadata.rawValue, flags:[], tx:someTrans)
-            let scheduledTasks = try makeEnv.openDatabase(named:Databases.scheduleTasks.rawValue, flags:[], tx:someTrans)
-            let scheduleIntervalDB = try makeEnv.openDatabase(named:Databases.scheduleInterval.rawValue, flags:[], tx:someTrans)
-            let scheduleLastFire = try makeEnv.openDatabase(named:Databases.scheduleLastFireDate.rawValue, flags:[], tx:someTrans)
-			let notifyDB = try makeEnv.openDatabase(named:Databases.notifyUsers.rawValue, flags:[], tx:someTrans)
-			
-            if running {
-                do {
-                    let lastPid = try metadataDB.getEntry(type:pid_t.self, forKey:Metadatas.daemonRunningPID.rawValue, tx:someTrans)!
-                    let checkPid = kill(lastPid, 0)
-                    guard checkPid != 0 else {
-                        throw Error.daemonAlreadyRunning
-                    }
-                } catch LMDBError.notFound {}
-                try metadataDB.setEntry(value:getpid(), forKey:Metadatas.daemonRunningPID.rawValue, tx:someTrans)
-                try scheduledTasks.deleteAllEntries(tx:someTrans)
-            }
+    init(running:Bool = true) throws {
+		let directory = URL(fileURLWithPath:"/var/lib/wiremand", isDirectory:true)
+		
+		// define the paths of the database
+		let makeEnvPath = directory.appendingPathComponent("daemon-dbi")
+
+		let ro:Bool
+		var mdb_flags:Environment.Flags = [.noSync, .noSubDir, .noReadAhead]
+		
+		//validate the file exists
+		if (access(makeEnvPath.path, F_OK | R_OK | X_OK) == 0) {
+			if access(makeEnvPath.path, W_OK) != 0 {
+				ro = true
+				mdb_flags.update(with:.readOnly)
+				Self.logger.trace("readonly mode enabled.")
+			} else {
+				ro = false
+				Self.logger.trace("read/write mode enabled.")
+			}
+		} else {
+			throw Error.databaseNotCreated
+		}
+		
+		let makeEnv:Environment
+		do {
+			makeEnv = try Environment(path:makeEnvPath.path, flags:mdb_flags, mapSize:5000000000, maxDBs:128)
+			Self.logger.trace("lmdb environment initialized.")
+		} catch let error {
+			Self.logger.error("unable to initialize lmdb environment.", metadata:["error":"\(error)"])
+			throw error
+		}
+		
+		let dbs = try makeEnv.transact(readOnly:ro) { someTrans -> [Database] in
+            let metadataDB = try makeEnv.openDatabase(named:Databases.metadata.rawValue, tx:someTrans)
+            let scheduledTasks = try makeEnv.openDatabase(named:Databases.scheduleTasks.rawValue, tx:someTrans)
+            let scheduleIntervalDB = try makeEnv.openDatabase(named:Databases.scheduleInterval.rawValue, tx:someTrans)
+            let scheduleLastFire = try makeEnv.openDatabase(named:Databases.scheduleLastFireDate.rawValue, tx:someTrans)
+			let notifyDB = try makeEnv.openDatabase(named:Databases.notifyUsers.rawValue, tx:someTrans)
+			if running {
+				do {
+					let lastPid = try metadataDB.getEntry(type:pid_t.self, forKey:Metadatas.daemonRunningPID.rawValue, tx:someTrans)!
+					let checkPid = kill(lastPid, 0)
+					guard checkPid != 0 else {
+						Self.logger.error("running mode enabled but daemon process already running", metadata:["this_pid":"\(getpid())", "running_pid":"\(lastPid)"])
+						throw Error.daemonAlreadyRunning
+					}
+				} catch LMDBError.notFound {}
+				Self.logger.trace("running mode enabled. committing PID in database.", metadata:["this_pid":"\(getpid())"])
+				try metadataDB.setEntry(value:getpid(), forKey:Metadatas.daemonRunningPID.rawValue, tx:someTrans)
+				try scheduledTasks.deleteAllEntries(tx:someTrans)
+			}
+			Self.logger.debug("instance initialized successfully.", metadata:["readonly": "\(ro)", "this_running_pid": "\(running)"])
             return [metadataDB, scheduledTasks, scheduleIntervalDB, scheduleLastFire, notifyDB]
         }
+		
         self.env = makeEnv
         self.metadata = dbs[0]
         self.scheduledTasks = dbs[1]
@@ -105,8 +161,17 @@ actor DaemonDB {
         self.scheduleLastFire = dbs[3]
 		self.notifyDB = dbs[4]
         self.wireguardDatabase = try WireguardDatabase(environment:makeEnv)
-		self.printerDatabase = try PrintDB(environment:makeEnv, directory:directory)
+		let makePrintDB:PrintDB?
+		do {
+			makePrintDB = try PrintDB(environment:makeEnv, directory:directory, readOnly:ro)
+		} catch LMDBError.notFound {
+			Self.logger.debug("print db not enabled")
+			makePrintDB = nil
+		}
+		self.printerDatabase = makePrintDB
 		self.ipdb = try IPDatabase.init(base:directory)
+		self.readOnly = ro
+		self.running = running
     }
     enum Schedule:String {
         case latestWireguardHandshakesCheck = "_wg_latestHandshakesCheck"
@@ -216,16 +281,18 @@ actor DaemonDB {
 	}
     
     deinit {
-        try! env.transact(readOnly:false) { someTrans in
-            let curPID = getpid()
-            do {
-                let checkPid = try metadata.getEntry(type:pid_t.self, forKey:Metadatas.daemonRunningPID.rawValue, tx:someTrans)!
-                if curPID == checkPid {
-                    try metadata.deleteEntry(key:checkPid, tx:someTrans)
-                    try scheduledTasks.deleteAllEntries(tx:someTrans)
-                }
-            } catch LMDBError.notFound {}
-        }
-		try! env.sync(force:true)
+		if (running == true && readOnly == false) {
+			try! env.transact(readOnly:false) { someTrans in
+				let curPID = getpid()
+				do {
+					let checkPid = try metadata.getEntry(type:pid_t.self, forKey:Metadatas.daemonRunningPID.rawValue, tx:someTrans)!
+					if curPID == checkPid {
+						try metadata.deleteEntry(key:checkPid, tx:someTrans)
+						try scheduledTasks.deleteAllEntries(tx:someTrans)
+					}
+				} catch LMDBError.notFound {}
+			}
+			try! env.sync(force:true)
+		}
     }
 }
